+++ docs/README.md
@@ -50,8 +50,9 @@
 
 # 1단계 리뷰 & 수정사항 정리
 
-- [ ] 도메인 패키지를 정리하자
+- [x] 도메인 패키지를 정리하자
     - 코드를 읽는데 많은 도움이 되고, 내가 생각한 도메인 구조를 이해하는데도 많은 도움이 된다.
+    - 카드와 관련된 패키지(`card`), 플레이어와 관련된 패키지(`player`)로 분리했다.
 - [ ] 컨벤션을 잘 지켰는지 전체적으로 다시 확인해보자
     - `isBLACKJACK()`처럼 완전히 틀린것
 - [ ] 이름이 의도를 잘 나타내는지 다시 확인해보자(클래스, 메서드, 변수 등). 명시적으로 변경해볼 수 있으면 해보자.
