+++ src/main/java/blackjack/domain/player/AbstractPlayer.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.Cards;
 import java.util.ArrayList;
 import java.util.Objects;
