+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.player;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
 
 public interface Player {
