+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.card.Deck;
 import java.util.List;
 import java.util.Objects;
