+++ src/test/java/blackjack/domain/card/FixDeck.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public class FixDeck implements Deck {
     @Override
